requirements.txt
python-telegram-bot~=20.7
gtts~=2.5.4
speechrecognition~=3.14.3
pydub~=0.25.1

joblib~=1.5.0
scikit-learn~=1.6.1
